moved ClubMember.cs to the apropriate directory
