+++ src/ClubService/ClubService.Domain/ClubAggregate/Entities/ClubMember.cs
@@ -1,7 +1,7 @@
 using ClubService.Domain.ClubAggregate.ValueObjects;
 using ClubService.Domain.Common.Abstract;
 
-namespace ClubService.Domain.ClubAggregate;
+namespace ClubService.Domain.ClubAggregate.Entities;
 
 public sealed class ClubMember : Entity<int>
 {
